\name{esnXmind-package}
\alias{esnXmind-package}
\alias{esnXmind}
\docType{package}
\title{
  Allows EsN to interact with R.
}
\description{
  Allows EsN to interact with R.
}
\details{
  \tabular{ll}{
    Package: \tab esnR\cr
    Type: \tab Package\cr
    Version: \tab 0.13\cr
    Date: \tab 2010-11-30\cr
    License: \tab GPL3\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Matthew Fidler

  Maintainer: Matthew Fidler <matthew.fidler@gmail.com>
}
\references{
}
\keyword{ EsN }
